+++ src/main/java/com/example/desafio_profesional_back/repositories/ProductoRepository.java
@@ -5,7 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 /**
  * Repository is an interface that provides access to data in a database
+ * Repositorio para operaciones CRUD con la entidad Producto.
  */
-public interface ProductoRepository extends JpaRepository<Producto, Integer> {
+public interface ProductoRepository extends JpaRepository<Producto, Long> {
     boolean existsByNombre(String nombre);
 }
